Question_no_1__if_else
Given an integer,n , perform the following conditional actions:

case 1 :If  is odd, print Weird.

case 2 :If  is even and in the inclusive range of 2 to 5, print Not Weird.

case 3 :If  is even and in the inclusive range of 6 to 20 , print Weird.

case 4 :If  is even and greater than 20, print Not Weird.

Complete the stub code provided in your editor to print whether or not n is weird.